web/requirements.txt
python-dotenv~=0.15.0
pytest~=6.2.1
asyncpg~=0.22.0
aiohttp~=3.7.3
requests~=2.25.1
beautifulsoup4~=4.9.1
sqlalchemy~=1.4.13
sqlalchemy-utils~=0.36.8
python-whois~=0.7.3
gunicorn~=20.1.0